refactor(screenshots): split workflow files by responsibility
