+++ internal/cli/assets/assets_screenshots.go
@@ -6,8 +6,6 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"path/filepath"
-	"sort"
 	"strings"
 
 	"github.com/peterbourgon/ff/v3/ffcli"
@@ -675,1364 +673,4 @@ func executeScreenshotUploadCommand(ctx context.Context, opts screenshotUploadCo
 		return &result, err
 	}
 	return &result, nil
-}
-
-func collectScreenshotUploadFiles(path string, maxScreenshots int) ([]string, error) {
-	files, err := collectAssetPaths(path)
-	if err != nil {
-		return nil, err
-	}
-	if len(files) == 0 {
-		return nil, fmt.Errorf("no files found in %q", path)
-	}
-
-	files, err = limitScreenshotUploadFiles(files, maxScreenshots, path)
-	if err != nil {
-		return nil, err
-	}
-	for _, filePath := range files {
-		if err := asc.ValidateImageFile(filePath); err != nil {
-			return nil, err
-		}
-	}
-	return files, nil
-}
-
-func limitScreenshotUploadFiles(files []string, maxScreenshots int, source string) ([]string, error) {
-	if maxScreenshots > 0 {
-		if len(files) > maxScreenshots {
-			return append([]string(nil), files[:maxScreenshots]...), nil
-		}
-		return files, nil
-	}
-
-	if len(files) > appScreenshotSetMaxScreenshots {
-		return nil, fmt.Errorf(
-			"%q contains %d screenshots for one screenshot set; App Store screenshot sets allow at most %d images. Remove extra files or pass --max-screenshots %d to upload the first %d sorted files",
-			source,
-			len(files),
-			appScreenshotSetMaxScreenshots,
-			appScreenshotSetMaxScreenshots,
-			appScreenshotSetMaxScreenshots,
-		)
-	}
-	return files, nil
-}
-
-func limitScreenshotFanoutUploadFiles(localeAssets []screenshotLocaleAssetFiles, maxScreenshots int) ([]screenshotLocaleAssetFiles, error) {
-	limited := make([]screenshotLocaleAssetFiles, 0, len(localeAssets))
-	for _, item := range localeAssets {
-		files, err := limitScreenshotUploadFiles(item.Files, maxScreenshots, item.Locale)
-		if err != nil {
-			return nil, err
-		}
-		item.Files = files
-		limited = append(limited, item)
-	}
-	return limited, nil
-}
-
-func limitScreenshotUploadFilesForExistingSet(files []string, maxScreenshots int, existingScreenshots []asc.Resource[asc.AppScreenshotAttributes], replace bool, setID string) ([]string, error) {
-	if replace {
-		if maxScreenshots <= 0 {
-			return files, nil
-		}
-		if len(files) > maxScreenshots {
-			return append([]string(nil), files[:maxScreenshots]...), nil
-		}
-		return files, nil
-	}
-
-	setLabel := strings.TrimSpace(setID)
-	if setLabel == "" {
-		setLabel = "target set"
-	}
-	if maxScreenshots <= 0 {
-		total := len(existingScreenshots) + len(files)
-		if total > appScreenshotSetMaxScreenshots {
-			return nil, fmt.Errorf(
-				"%s already has %d screenshot(s); uploading %d more would exceed App Store screenshot set limit %d. Pass --replace to replace existing screenshots or --max-screenshots %d to upload only the remaining slot(s)",
-				setLabel,
-				len(existingScreenshots),
-				len(files),
-				appScreenshotSetMaxScreenshots,
-				max(0, appScreenshotSetMaxScreenshots-len(existingScreenshots)),
-			)
-		}
-		return files, nil
-	}
-
-	remaining := maxScreenshots - len(existingScreenshots)
-	if remaining <= 0 {
-		if len(files) == 0 {
-			return files, nil
-		}
-		return nil, fmt.Errorf(
-			"%s already has %d screenshot(s); --max-screenshots %d leaves no upload slots. Pass --replace to replace existing screenshots or choose a higher limit up to %d",
-			setLabel,
-			len(existingScreenshots),
-			maxScreenshots,
-			appScreenshotSetMaxScreenshots,
-		)
-	}
-	if len(files) > remaining {
-		return append([]string(nil), files[:remaining]...), nil
-	}
-	return files, nil
-}
-
-func uploadScreenshotsFanout(ctx context.Context, cfg screenshotUploadFanoutConfig) (asc.AppScreenshotFanoutUploadResult, error) {
-	var zero asc.AppScreenshotFanoutUploadResult
-
-	if cfg.Client == nil {
-		return zero, fmt.Errorf("client is required")
-	}
-	if cfg.RequestContext == nil {
-		cfg.RequestContext = shared.ContextWithTimeout
-	}
-	cfg.ExecuteUpload = resolveScreenshotUploadExecutor(cfg.ExecuteUpload, cfg.UploadScreenshot)
-
-	localeAssets := cfg.LocaleAssets
-	var err error
-	if localeAssets == nil {
-		localeAssets, err = collectLocaleAssetFiles(cfg.RootPath, cfg.DisplayType)
-		if err != nil {
-			return zero, err
-		}
-		cfg.LocaleAssetsCanonical = true
-	}
-	if !cfg.LocaleAssetsCanonical {
-		localeAssets, err = canonicalizeUniqueScreenshotFanoutLocaleAssets(localeAssets)
-		if err != nil {
-			return zero, err
-		}
-	}
-
-	requestCtx, cancel := cfg.RequestContext(ctx)
-	localizationsResp, err := cfg.Client.GetAppStoreVersionLocalizations(requestCtx, cfg.VersionID, asc.WithAppStoreVersionLocalizationsLimit(200))
-	cancel()
-	if err != nil {
-		return zero, fmt.Errorf("fetch version localizations: %w", err)
-	}
-
-	localizationIDsByLocale := make(map[string]string, len(localizationsResp.Data))
-	for _, item := range localizationsResp.Data {
-		localeKey := normalizeFanoutLocaleKey(item.Attributes.Locale)
-		if localeKey == "" {
-			continue
-		}
-		localizationIDsByLocale[localeKey] = strings.TrimSpace(item.ID)
-	}
-
-	missingLocales := make([]string, 0)
-	for _, item := range localeAssets {
-		if localizationIDsByLocale[normalizeFanoutLocaleKey(item.Locale)] == "" {
-			missingLocales = append(missingLocales, item.Locale)
-		}
-	}
-	if len(missingLocales) > 0 {
-		sort.Strings(missingLocales)
-		return zero, fmt.Errorf("no matching App Store version localizations found for locales: %s", strings.Join(missingLocales, ", "))
-	}
-
-	result := asc.AppScreenshotFanoutUploadResult{
-		AppID:         cfg.AppID,
-		Version:       cfg.Version,
-		VersionID:     cfg.VersionID,
-		Platform:      cfg.Platform,
-		DisplayType:   cfg.DisplayType,
-		DryRun:        cfg.DryRun,
-		Localizations: make([]asc.AppScreenshotLocalizationUploadResult, 0, len(localeAssets)),
-	}
-
-	for _, item := range localeAssets {
-		localizationID := localizationIDsByLocale[normalizeFanoutLocaleKey(item.Locale)]
-		uploadResult, err := cfg.ExecuteUpload(ctx, screenshotUploadConfig[asc.AppScreenshotUploadResult]{
-			Client:         cfg.Client,
-			LocalizationID: localizationID,
-			DisplayType:    cfg.DisplayType,
-			Files:          item.Files,
-			SkipExisting:   cfg.SkipExisting,
-			Replace:        cfg.Replace,
-			DryRun:         cfg.DryRun,
-			MaxScreenshots: cfg.MaxScreenshots,
-			RequestContext: cfg.RequestContext,
-			UploadContext:  contextWithAssetUploadTimeout,
-			Access:         appStoreVersionScreenshotSetAccess,
-		}, "")
-		if err != nil {
-			if hasAppScreenshotUploadResultOutput(uploadResult) {
-				result.Localizations = append(result.Localizations, buildFanoutLocalizationUploadResult(item.Locale, uploadResult))
-			}
-			return result, fmt.Errorf("upload locale %s: %w", item.Locale, err)
-		}
-		result.Localizations = append(result.Localizations, buildFanoutLocalizationUploadResult(item.Locale, uploadResult))
-	}
-
-	return result, nil
-}
-
-func collectLocaleAssetFiles(rootPath, displayType string) ([]screenshotLocaleAssetFiles, error) {
-	return collectLocaleAssetFilesWithLimit(rootPath, displayType, 0)
-}
-
-func collectLocaleAssetFilesWithLimit(rootPath, displayType string, maxScreenshots int) ([]screenshotLocaleAssetFiles, error) {
-	info, err := os.Lstat(rootPath)
-	if err != nil {
-		return nil, err
-	}
-	if info.Mode()&os.ModeSymlink != 0 {
-		return nil, fmt.Errorf("refusing to read symlink %q", rootPath)
-	}
-	if !info.IsDir() {
-		return nil, fmt.Errorf("fan-out upload path %q must be a directory containing locale subdirectories", rootPath)
-	}
-
-	entries, err := os.ReadDir(rootPath)
-	if err != nil {
-		return nil, err
-	}
-
-	results := make([]screenshotLocaleAssetFiles, 0, len(entries))
-	seenLocales := make(map[string]string, len(entries))
-	for _, entry := range entries {
-		if shouldIgnoreFanoutEntryName(entry.Name()) {
-			continue
-		}
-
-		entryPath := filepath.Join(rootPath, entry.Name())
-		info, err := os.Lstat(entryPath)
-		if err != nil {
-			return nil, err
-		}
-		if info.Mode()&os.ModeSymlink != 0 {
-			return nil, fmt.Errorf("refusing to read symlink %q", entryPath)
-		}
-		if !info.IsDir() {
-			continue
-		}
-
-		locale, err := shared.CanonicalizeAppStoreLocalizationLocale(entry.Name())
-		if err != nil {
-			hasMatchingFiles, matchErr := directoryContainsMatchingScreenshotFiles(entryPath, displayType)
-			if matchErr != nil {
-				return nil, matchErr
-			}
-			if !hasMatchingFiles {
-				continue
-			}
-			return nil, fmt.Errorf("invalid locale directory %q: %w", entry.Name(), err)
-		}
-		if !isKnownAppStoreLocalizationLocale(locale) {
-			hasMatchingFiles, matchErr := directoryContainsMatchingScreenshotFiles(entryPath, displayType)
-			if matchErr != nil {
-				return nil, matchErr
-			}
-			if !hasMatchingFiles {
-				continue
-			}
-		}
-		files, err := collectLocaleAssetFilesRecursiveWithLimit(entryPath, displayType, maxScreenshots)
-		if err != nil {
-			return nil, fmt.Errorf("locale %s: %w", locale, err)
-		}
-		if err := registerUniqueCanonicalFanoutLocale(locale, entry.Name(), "fan-out path", "dirs", seenLocales); err != nil {
-			return nil, err
-		}
-		results = append(results, screenshotLocaleAssetFiles{
-			Locale: locale,
-			Files:  files,
-		})
-	}
-
-	if len(results) == 0 {
-		return nil, fmt.Errorf("no locale directories found in %q", rootPath)
-	}
-
-	sort.Slice(results, func(i, j int) bool {
-		return strings.ToLower(results[i].Locale) < strings.ToLower(results[j].Locale)
-	})
-	return results, nil
-}
-
-type screenshotMatchWalkOptions struct {
-	ignoreInvalidFiles bool
-	ignoreSymlinks     bool
-	onMatch            func(path string) error
-}
-
-func walkMatchingScreenshotFiles(rootPath, displayType string, opts screenshotMatchWalkOptions) error {
-	return filepath.WalkDir(rootPath, func(path string, entry os.DirEntry, walkErr error) error {
-		if walkErr != nil {
-			return walkErr
-		}
-		if path != rootPath && shouldIgnoreFanoutEntryName(entry.Name()) {
-			if entry.IsDir() {
-				return filepath.SkipDir
-			}
-			return nil
-		}
-		if entry.Type()&os.ModeSymlink != 0 {
-			if opts.ignoreSymlinks {
-				return nil
-			}
-			return fmt.Errorf("refusing to read symlink %q", path)
-		}
-		if entry.IsDir() {
-			return nil
-		}
-
-		info, err := entry.Info()
-		if err != nil {
-			return err
-		}
-		if !info.Mode().IsRegular() || !isSupportedScreenshotUploadFile(path) {
-			return nil
-		}
-		if err := asc.ValidateImageFile(path); err != nil {
-			if opts.ignoreInvalidFiles {
-				return nil
-			}
-			return err
-		}
-		isMatch, err := screenshotMatchesDisplayType(path, displayType)
-		if err != nil {
-			if opts.ignoreInvalidFiles {
-				return nil
-			}
-			return err
-		}
-		if !isMatch || opts.onMatch == nil {
-			return nil
-		}
-		if err := opts.onMatch(path); err != nil {
-			return err
-		}
-		return nil
-	})
-}
-
-func collectLocaleAssetFilesRecursive(rootPath, displayType string) ([]string, error) {
-	return collectLocaleAssetFilesRecursiveWithLimit(rootPath, displayType, 0)
-}
-
-func collectLocaleAssetFilesRecursiveWithLimit(rootPath, displayType string, maxScreenshots int) ([]string, error) {
-	if maxScreenshots > 0 {
-		return collectLimitedLocaleAssetFilesRecursive(rootPath, displayType, maxScreenshots)
-	}
-
-	files := make([]string, 0)
-	err := walkMatchingScreenshotFiles(rootPath, displayType, screenshotMatchWalkOptions{
-		onMatch: func(path string) error {
-			files = append(files, path)
-			return nil
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-	if len(files) == 0 {
-		return nil, fmt.Errorf("no screenshot files matching %s found in %q", displayType, rootPath)
-	}
-	sort.Strings(files)
-	if err := validateUniqueScreenshotUploadFileNames(files); err != nil {
-		return nil, err
-	}
-	return files, nil
-}
-
-func collectLimitedLocaleAssetFilesRecursive(rootPath, displayType string, maxScreenshots int) ([]string, error) {
-	candidates, err := collectSupportedScreenshotCandidateFiles(rootPath)
-	if err != nil {
-		return nil, err
-	}
-
-	files := make([]string, 0, min(maxScreenshots, len(candidates)))
-	for _, path := range candidates {
-		if len(files) >= maxScreenshots {
-			break
-		}
-		if err := asc.ValidateImageFile(path); err != nil {
-			return nil, err
-		}
-		matches, err := screenshotMatchesDisplayType(path, displayType)
-		if err != nil {
-			return nil, err
-		}
-		if matches {
-			files = append(files, path)
-		}
-	}
-	if len(files) == 0 {
-		return nil, fmt.Errorf("no screenshot files matching %s found in %q", displayType, rootPath)
-	}
-	if err := validateUniqueScreenshotUploadFileNames(files); err != nil {
-		return nil, err
-	}
-	return files, nil
-}
-
-func validateUniqueScreenshotUploadFileNames(files []string) error {
-	seen := make(map[string]string, len(files))
-	for _, filePath := range files {
-		fileName := filepath.Base(filePath)
-		key := strings.ToLower(fileName)
-		if previous, ok := seen[key]; ok {
-			return fmt.Errorf(
-				"duplicate screenshot file name %q (%q, %q); App Store screenshot ordering uses uploaded file names, so rename one file before uploading",
-				fileName,
-				previous,
-				filePath,
-			)
-		}
-		seen[key] = filePath
-	}
-	return nil
-}
-
-func collectSupportedScreenshotCandidateFiles(rootPath string) ([]string, error) {
-	files := make([]string, 0)
-	err := filepath.WalkDir(rootPath, func(path string, entry os.DirEntry, walkErr error) error {
-		if walkErr != nil {
-			return walkErr
-		}
-		if path != rootPath && shouldIgnoreFanoutEntryName(entry.Name()) {
-			if entry.IsDir() {
-				return filepath.SkipDir
-			}
-			return nil
-		}
-		if entry.Type()&os.ModeSymlink != 0 {
-			return fmt.Errorf("refusing to read symlink %q", path)
-		}
-		if entry.IsDir() {
-			return nil
-		}
-		info, err := entry.Info()
-		if err != nil {
-			return err
-		}
-		if !info.Mode().IsRegular() || !isSupportedScreenshotUploadFile(path) {
-			return nil
-		}
-		files = append(files, path)
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	sort.Strings(files)
-	return files, nil
-}
-
-func directoryContainsMatchingScreenshotFiles(rootPath, displayType string) (bool, error) {
-	found := false
-	err := walkMatchingScreenshotFiles(rootPath, displayType, screenshotMatchWalkOptions{
-		ignoreInvalidFiles: true,
-		ignoreSymlinks:     true,
-		onMatch: func(path string) error {
-			found = true
-			return filepath.SkipAll
-		},
-	})
-	if err != nil && !errors.Is(err, filepath.SkipAll) {
-		return false, err
-	}
-	return found, nil
-}
-
-func canonicalizeUniqueScreenshotFanoutLocaleAssets(localeAssets []screenshotLocaleAssetFiles) ([]screenshotLocaleAssetFiles, error) {
-	result := make([]screenshotLocaleAssetFiles, 0, len(localeAssets))
-	seen := make(map[string]string, len(localeAssets))
-	for _, item := range localeAssets {
-		canonicalLocale, err := shared.CanonicalizeAppStoreLocalizationLocale(item.Locale)
-		if err != nil {
-			return nil, fmt.Errorf("invalid locale %q in fan-out upload: %w", item.Locale, err)
-		}
-		if err := registerUniqueCanonicalFanoutLocale(canonicalLocale, item.Locale, "fan-out upload", "inputs", seen); err != nil {
-			return nil, err
-		}
-		if err := validateUniqueScreenshotUploadFileNames(item.Files); err != nil {
-			return nil, fmt.Errorf("locale %s: %w", canonicalLocale, err)
-		}
-		result = append(result, screenshotLocaleAssetFiles{
-			Locale: canonicalLocale,
-			Files:  item.Files,
-		})
-	}
-	return result, nil
-}
-
-func registerUniqueCanonicalFanoutLocale(canonicalLocale, source, scope, itemLabel string, seen map[string]string) error {
-	localeKey := normalizeFanoutLocaleKey(canonicalLocale)
-	if previous, ok := seen[localeKey]; ok {
-		return fmt.Errorf("duplicate locale %q in %s (%s: %q, %q)", canonicalLocale, scope, itemLabel, previous, source)
-	}
-	seen[localeKey] = source
-	return nil
-}
-
-func normalizeFanoutLocaleKey(locale string) string {
-	return strings.ToLower(shared.NormalizeLocaleCode(locale))
-}
-
-func isKnownAppStoreLocalizationLocale(locale string) bool {
-	_, ok := knownAppStoreLocalizationLocales[normalizeFanoutLocaleKey(locale)]
-	return ok
-}
-
-func screenshotMatchesDisplayType(path, displayType string) (bool, error) {
-	allowed, ok := asc.ScreenshotDimensions(displayType)
-	if !ok {
-		return false, fmt.Errorf("unsupported screenshot display type %q", displayType)
-	}
-
-	dims, err := asc.ReadImageDimensions(path)
-	if err != nil {
-		return false, err
-	}
-
-	for _, dim := range allowed {
-		if dim.Width == dims.Width && dim.Height == dims.Height {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-func shouldIgnoreFanoutEntryName(name string) bool {
-	return strings.HasPrefix(name, ".")
-}
-
-func isSupportedScreenshotUploadFile(path string) bool {
-	switch strings.ToLower(filepath.Ext(path)) {
-	case ".png", ".jpg", ".jpeg":
-		return true
-	default:
-		return false
-	}
-}
-
-type screenshotDownloadItem struct {
-	ID          string `json:"id"`
-	DisplayType string `json:"displayType,omitempty"`
-	FileName    string `json:"fileName,omitempty"`
-	URL         string `json:"url,omitempty"`
-	OutputPath  string `json:"outputPath"`
-
-	ContentType  string `json:"contentType,omitempty"`
-	BytesWritten int64  `json:"bytesWritten,omitempty"`
-}
-
-type screenshotDownloadFailure struct {
-	ID          string `json:"id,omitempty"`
-	DisplayType string `json:"displayType,omitempty"`
-	URL         string `json:"url,omitempty"`
-	OutputPath  string `json:"outputPath,omitempty"`
-	Error       string `json:"error"`
-}
-
-type screenshotDownloadResult struct {
-	VersionLocalizationID string `json:"versionLocalizationId,omitempty"`
-	OutputDir             string `json:"outputDir,omitempty"`
-	Overwrite             bool   `json:"overwrite"`
-
-	Total      int `json:"total"`
-	Downloaded int `json:"downloaded"`
-	Failed     int `json:"failed"`
-
-	Items    []screenshotDownloadItem    `json:"items,omitempty"`
-	Failures []screenshotDownloadFailure `json:"failures,omitempty"`
-}
-
-// AssetsScreenshotsDownloadCommand returns the screenshots download subcommand.
-func AssetsScreenshotsDownloadCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("download", flag.ExitOnError)
-
-	id := fs.String("id", "", "Screenshot ID to download")
-	localizationID := fs.String("version-localization", "", "App Store version localization ID (download all screenshots)")
-	outputPath := fs.String("output", "", "Output file path (required with --id)")
-	outputDir := fs.String("output-dir", "", "Output directory (required with --version-localization)")
-	overwrite := fs.Bool("overwrite", false, "Overwrite existing files")
-	format := shared.BindOutputFlagsWith(fs, "format", "json", "Summary output format: json (default), table, markdown")
-
-	return &ffcli.Command{
-		Name:       "download",
-		ShortUsage: "asc screenshots download (--id \"SCREENSHOT_ID\" --output \"./screenshot.png\") | (--version-localization \"VERSION_LOCALIZATION_ID\" --output-dir \"./screenshots\")",
-		ShortHelp:  "Download App Store screenshots to disk.",
-		LongHelp: `Download App Store screenshots to disk.
-
---version-localization is the App Store version localization resource ID
-returned as data[].id by "asc localizations list --version VERSION_ID --output json".
-It is not the locale code such as en-US.
-
-Examples:
-  asc screenshots download --id "SCREENSHOT_ID" --output "./screenshot.png"
-  asc localizations list --version "VERSION_ID" --output json --locale "en-US"
-  asc screenshots download --version-localization "VERSION_LOCALIZATION_ID" --output-dir "./screenshots"
-  asc screenshots download --version-localization "VERSION_LOCALIZATION_ID" --output-dir "./screenshots" --overwrite`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			idValue := strings.TrimSpace(*id)
-			locID := strings.TrimSpace(*localizationID)
-
-			if idValue == "" && locID == "" {
-				fmt.Fprintln(os.Stderr, "Error: --id or --version-localization is required")
-				return shared.MissingRequiredUsageError()
-			}
-			if idValue != "" && locID != "" {
-				return shared.UsageError("--id and --version-localization are mutually exclusive")
-			}
-
-			outputFile := strings.TrimSpace(*outputPath)
-			outputDirValue := strings.TrimSpace(*outputDir)
-			if idValue != "" {
-				if outputFile == "" {
-					fmt.Fprintln(os.Stderr, "Error: --output is required with --id")
-					return shared.MissingRequiredUsageError()
-				}
-				if strings.HasSuffix(outputFile, string(filepath.Separator)) {
-					return shared.UsageError("--output must be a file path")
-				}
-			}
-			if locID != "" {
-				if outputDirValue == "" {
-					fmt.Fprintln(os.Stderr, "Error: --output-dir is required with --version-localization")
-					return shared.MissingRequiredUsageError()
-				}
-			}
-
-			client, err := shared.GetASCClient()
-			if err != nil {
-				return fmt.Errorf("screenshots download: %w", err)
-			}
-
-			cleanOutputDir := ""
-			if outputDirValue != "" {
-				cleanOutputDir = filepath.Clean(outputDirValue)
-			}
-			result := &screenshotDownloadResult{
-				VersionLocalizationID: locID,
-				OutputDir:             cleanOutputDir,
-				Overwrite:             *overwrite,
-			}
-
-			items := make([]screenshotDownloadItem, 0, 8)
-
-			if idValue != "" {
-				requestCtx, cancel := shared.ContextWithTimeout(ctx)
-				resp, err := client.GetAppScreenshot(requestCtx, idValue)
-				cancel()
-				if err != nil {
-					return fmt.Errorf("screenshots download: failed to fetch screenshot: %w", err)
-				}
-
-				downloadURL, err := resolveImageAssetDownloadURL(resp.Data.Attributes.ImageAsset, resp.Data.Attributes.FileName)
-				if err != nil {
-					items = append(items, screenshotDownloadItem{
-						ID:         idValue,
-						FileName:   strings.TrimSpace(resp.Data.Attributes.FileName),
-						OutputPath: outputFile,
-					})
-					result.Items = items
-					result.Failures = append(result.Failures, screenshotDownloadFailure{
-						ID:         idValue,
-						OutputPath: outputFile,
-						Error:      err.Error(),
-					})
-					result.Total = 1
-					result.Failed = 1
-
-					if err := shared.PrintOutputWithRenderers(
-						result,
-						*format.Output,
-						*format.Pretty,
-						func() error { return renderScreenshotDownloadResult(result, false) },
-						func() error { return renderScreenshotDownloadResult(result, true) },
-					); err != nil {
-						return err
-					}
-					return shared.NewReportedError(fmt.Errorf("screenshots download: 1 file failed"))
-				}
-
-				items = append(items, screenshotDownloadItem{
-					ID:         idValue,
-					FileName:   strings.TrimSpace(resp.Data.Attributes.FileName),
-					URL:        downloadURL,
-					OutputPath: outputFile,
-				})
-			} else {
-				requestCtx, cancel := shared.ContextWithTimeout(ctx)
-				setsResp, err := client.GetAppScreenshotSets(requestCtx, locID)
-				cancel()
-				if err != nil {
-					return fmt.Errorf("screenshots download: failed to fetch sets: %w", err)
-				}
-
-				sets := make([]asc.Resource[asc.AppScreenshotSetAttributes], 0, len(setsResp.Data))
-				sets = append(sets, setsResp.Data...)
-				sort.Slice(sets, func(i, j int) bool {
-					di := strings.ToUpper(strings.TrimSpace(sets[i].Attributes.ScreenshotDisplayType))
-					dj := strings.ToUpper(strings.TrimSpace(sets[j].Attributes.ScreenshotDisplayType))
-					if di == dj {
-						return sets[i].ID < sets[j].ID
-					}
-					return di < dj
-				})
-
-				for _, set := range sets {
-					displayType := strings.TrimSpace(set.Attributes.ScreenshotDisplayType)
-
-					requestCtx, cancel := shared.ContextWithTimeout(ctx)
-					shotsResp, err := client.GetAppScreenshots(requestCtx, set.ID)
-					cancel()
-					if err != nil {
-						return fmt.Errorf("screenshots download: failed to fetch screenshots for set %s: %w", set.ID, err)
-					}
-
-					requestCtx, cancel = shared.ContextWithTimeout(ctx)
-					orderedIDs, err := GetOrderedAppScreenshotIDs(requestCtx, client, set.ID)
-					cancel()
-					if err != nil {
-						return fmt.Errorf("screenshots download: failed to fetch screenshot order for set %s: %w", set.ID, err)
-					}
-					shots := orderScreenshotsForDownload(shotsResp.Data, orderedIDs)
-
-					for idx, shot := range shots {
-						base := sanitizeBaseFileName(shot.Attributes.FileName)
-						if base == "" {
-							base = strings.TrimSpace(shot.ID)
-						}
-						if base == "" {
-							base = fmt.Sprintf("screenshot-%d", idx+1)
-						}
-
-						destDir := filepath.Join(outputDirValue, displayType)
-						destName := fmt.Sprintf("%02d_%s_%s", idx+1, strings.TrimSpace(shot.ID), base)
-						destPath := filepath.Join(destDir, destName)
-
-						imageAsset := shot.Attributes.ImageAsset
-						if imageAsset == nil || strings.TrimSpace(imageAsset.TemplateURL) == "" {
-							requestCtx, cancel := shared.ContextWithTimeout(ctx)
-							full, err := client.GetAppScreenshot(requestCtx, shot.ID)
-							cancel()
-							if err == nil {
-								imageAsset = full.Data.Attributes.ImageAsset
-							}
-						}
-
-						downloadURL, err := resolveImageAssetDownloadURL(imageAsset, shot.Attributes.FileName)
-						if err != nil {
-							items = append(items, screenshotDownloadItem{
-								ID:          strings.TrimSpace(shot.ID),
-								DisplayType: displayType,
-								FileName:    strings.TrimSpace(shot.Attributes.FileName),
-								OutputPath:  destPath,
-							})
-							result.Failures = append(result.Failures, screenshotDownloadFailure{
-								ID:          strings.TrimSpace(shot.ID),
-								DisplayType: displayType,
-								OutputPath:  destPath,
-								Error:       err.Error(),
-							})
-							continue
-						}
-
-						items = append(items, screenshotDownloadItem{
-							ID:          strings.TrimSpace(shot.ID),
-							DisplayType: displayType,
-							FileName:    strings.TrimSpace(shot.Attributes.FileName),
-							URL:         downloadURL,
-							OutputPath:  destPath,
-						})
-					}
-				}
-			}
-
-			for i := range items {
-				item := &items[i]
-				if strings.TrimSpace(item.URL) == "" {
-					continue
-				}
-
-				downloadCtx, cancel := shared.ContextWithTimeout(ctx)
-				written, contentType, err := downloadURLToFile(downloadCtx, item.URL, item.OutputPath, *overwrite)
-				cancel()
-				if err != nil {
-					result.Failures = append(result.Failures, screenshotDownloadFailure{
-						ID:          item.ID,
-						DisplayType: item.DisplayType,
-						URL:         item.URL,
-						OutputPath:  item.OutputPath,
-						Error:       err.Error(),
-					})
-					continue
-				}
-
-				item.BytesWritten = written
-				item.ContentType = contentType
-				result.Downloaded++
-			}
-
-			result.Items = items
-			result.Total = len(items)
-			result.Failed = len(result.Failures)
-
-			if err := shared.PrintOutputWithRenderers(
-				result,
-				*format.Output,
-				*format.Pretty,
-				func() error { return renderScreenshotDownloadResult(result, false) },
-				func() error { return renderScreenshotDownloadResult(result, true) },
-			); err != nil {
-				return err
-			}
-
-			if result.Failed > 0 {
-				return shared.NewReportedError(fmt.Errorf("screenshots download: %d file(s) failed", result.Failed))
-			}
-			return nil
-		},
-	}
-}
-
-func orderScreenshotsForDownload(shots []asc.Resource[asc.AppScreenshotAttributes], orderedIDs []string) []asc.Resource[asc.AppScreenshotAttributes] {
-	ordered := append([]asc.Resource[asc.AppScreenshotAttributes](nil), shots...)
-	orderByID := make(map[string]int, len(orderedIDs))
-	for idx, id := range orderedIDs {
-		id = strings.TrimSpace(id)
-		if id == "" {
-			continue
-		}
-		if _, exists := orderByID[id]; exists {
-			continue
-		}
-		orderByID[id] = idx
-	}
-
-	sort.SliceStable(ordered, func(i, j int) bool {
-		iOrder, iOK := orderByID[strings.TrimSpace(ordered[i].ID)]
-		jOrder, jOK := orderByID[strings.TrimSpace(ordered[j].ID)]
-		switch {
-		case iOK && jOK:
-			return iOrder < jOrder
-		case iOK:
-			return true
-		case jOK:
-			return false
-		}
-
-		fi := strings.ToLower(strings.TrimSpace(ordered[i].Attributes.FileName))
-		fj := strings.ToLower(strings.TrimSpace(ordered[j].Attributes.FileName))
-		if fi == fj {
-			return ordered[i].ID < ordered[j].ID
-		}
-		return fi < fj
-	})
-
-	return ordered
-}
-
-func renderScreenshotDownloadResult(result *screenshotDownloadResult, markdown bool) error {
-	if result == nil {
-		return fmt.Errorf("result is nil")
-	}
-
-	render := asc.RenderTable
-	if markdown {
-		render = asc.RenderMarkdown
-	}
-
-	render(
-		[]string{"Version Localization", "Output Dir", "Overwrite", "Total", "Downloaded", "Failed"},
-		[][]string{{
-			result.VersionLocalizationID,
-			result.OutputDir,
-			fmt.Sprintf("%t", result.Overwrite),
-			fmt.Sprintf("%d", result.Total),
-			fmt.Sprintf("%d", result.Downloaded),
-			fmt.Sprintf("%d", result.Failed),
-		}},
-	)
-
-	if len(result.Items) > 0 {
-		rows := make([][]string, 0, len(result.Items))
-		for _, item := range result.Items {
-			rows = append(rows, []string{
-				item.ID,
-				item.DisplayType,
-				item.FileName,
-				item.OutputPath,
-				fmt.Sprintf("%d", item.BytesWritten),
-			})
-		}
-		render([]string{"ID", "Display Type", "File Name", "Output Path", "Bytes"}, rows)
-	}
-
-	if len(result.Failures) > 0 {
-		rows := make([][]string, 0, len(result.Failures))
-		for _, f := range result.Failures {
-			rows = append(rows, []string{
-				f.ID,
-				f.DisplayType,
-				f.OutputPath,
-				f.Error,
-			})
-		}
-		render([]string{"ID", "Display Type", "Output Path", "Error"}, rows)
-	}
-
-	return nil
-}
-
-// AssetsScreenshotsDeleteCommand returns the screenshot delete subcommand.
-func AssetsScreenshotsDeleteCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("delete", flag.ExitOnError)
-
-	id := fs.String("id", "", "Screenshot ID")
-	confirm := fs.Bool("confirm", false, "Confirm deletion")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "delete",
-		ShortUsage: "asc screenshots delete --id \"SCREENSHOT_ID\" --confirm",
-		ShortHelp:  "Delete a screenshot by ID.",
-		LongHelp: `Delete a screenshot by ID.
-
-Examples:
-  asc screenshots delete --id "SCREENSHOT_ID" --confirm`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			assetID := strings.TrimSpace(*id)
-			if assetID == "" {
-				fmt.Fprintln(os.Stderr, "Error: --id is required")
-				return shared.MissingRequiredUsageError()
-			}
-			if !*confirm {
-				fmt.Fprintln(os.Stderr, "Error: --confirm is required to delete")
-				return shared.MissingRequiredUsageError()
-			}
-
-			client, err := shared.GetASCClient()
-			if err != nil {
-				return fmt.Errorf("screenshots delete: %w", err)
-			}
-
-			requestCtx, cancel := shared.ContextWithTimeout(ctx)
-			defer cancel()
-
-			if err := client.DeleteAppScreenshot(requestCtx, assetID); err != nil {
-				return fmt.Errorf("screenshots delete: %w", err)
-			}
-
-			result := asc.AssetDeleteResult{
-				ID:      assetID,
-				Deleted: true,
-			}
-
-			return shared.PrintOutput(&result, *output.Output, *output.Pretty)
-		},
-	}
-}
-
-func normalizeScreenshotDisplayType(input string) (string, error) {
-	value := strings.ToUpper(strings.TrimSpace(input))
-	if value == "" {
-		return "", fmt.Errorf("device type is required")
-	}
-	if !strings.HasPrefix(value, "APP_") && !strings.HasPrefix(value, "IMESSAGE_") {
-		value = "APP_" + value
-	}
-	value = normalizeScreenshotDisplayTypeAlias(value)
-	if !asc.IsValidScreenshotDisplayType(value) {
-		return "", fmt.Errorf("unsupported screenshot display type %q", value)
-	}
-	return value, nil
-}
-
-// NormalizeScreenshotDisplayType normalizes and validates a screenshot display type.
-func NormalizeScreenshotDisplayType(input string) (string, error) {
-	return normalizeScreenshotDisplayType(input)
-}
-
-func normalizeScreenshotDisplayTypeAlias(value string) string {
-	return value
-}
-
-func validateScreenshotDimensions(files []string, displayType string) error {
-	for _, filePath := range files {
-		if err := asc.ValidateScreenshotDimensions(filePath, displayType); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// ValidateScreenshotDimensions validates screenshot dimensions for all files.
-func ValidateScreenshotDimensions(files []string, displayType string) error {
-	return validateScreenshotDimensions(files, displayType)
-}
-
-func uploadScreenshots(ctx context.Context, client *asc.Client, localizationID, displayType string, files []string, skipExisting, replace, dryRun bool) (asc.AppScreenshotUploadResult, error) {
-	result, err := uploadScreenshotsWithConfig(ctx, screenshotUploadConfig[asc.AppScreenshotUploadResult]{
-		Client:         client,
-		LocalizationID: localizationID,
-		DisplayType:    displayType,
-		Files:          files,
-		SkipExisting:   skipExisting,
-		Replace:        replace,
-		DryRun:         dryRun,
-		RequestContext: shared.ContextWithTimeout,
-		UploadContext:  contextWithAssetUploadTimeout,
-		Access:         appStoreVersionScreenshotSetAccess,
-		BuildResult: func(localizationID string, set asc.Resource[asc.AppScreenshotSetAttributes], dryRun bool, results []asc.AssetUploadResultItem) asc.AppScreenshotUploadResult {
-			return buildAppScreenshotUploadResult(localizationID, set, dryRun, results)
-		},
-	})
-	return result, err
-}
-
-func findScreenshotSetWithAccess(ctx context.Context, client *asc.Client, localizationID, displayType string, access ScreenshotSetAccess) (asc.Resource[asc.AppScreenshotSetAttributes], error) {
-	if access.List == nil {
-		return asc.Resource[asc.AppScreenshotSetAttributes]{}, fmt.Errorf("screenshot set list function is required")
-	}
-
-	resp, err := access.List(ctx, client, localizationID)
-	if err != nil {
-		return asc.Resource[asc.AppScreenshotSetAttributes]{}, err
-	}
-	for _, set := range resp.Data {
-		if strings.EqualFold(set.Attributes.ScreenshotDisplayType, displayType) {
-			return set, nil
-		}
-	}
-	return asc.Resource[asc.AppScreenshotSetAttributes]{
-		Attributes: asc.AppScreenshotSetAttributes{ScreenshotDisplayType: displayType},
-	}, nil
-}
-
-func ensureScreenshotSetWithAccess(ctx context.Context, client *asc.Client, localizationID, displayType string, access ScreenshotSetAccess) (asc.Resource[asc.AppScreenshotSetAttributes], error) {
-	if access.Create == nil {
-		return asc.Resource[asc.AppScreenshotSetAttributes]{}, fmt.Errorf("screenshot set create function is required")
-	}
-
-	set, err := findScreenshotSetWithAccess(ctx, client, localizationID, displayType, access)
-	if err != nil {
-		return asc.Resource[asc.AppScreenshotSetAttributes]{}, err
-	}
-	if set.ID != "" {
-		return set, nil
-	}
-
-	created, err := access.Create(ctx, client, localizationID, displayType)
-	if err != nil {
-		return asc.Resource[asc.AppScreenshotSetAttributes]{}, err
-	}
-	return created.Data, nil
-}
-
-func uploadScreenshotsWithConfig[T any](ctx context.Context, cfg screenshotUploadConfig[T]) (T, error) {
-	var zero T
-
-	if cfg.Client == nil {
-		return zero, fmt.Errorf("client is required")
-	}
-	if cfg.BuildResult == nil {
-		return zero, fmt.Errorf("build result function is required")
-	}
-	if cfg.RequestContext == nil {
-		cfg.RequestContext = shared.ContextWithTimeout
-	}
-	if cfg.UploadContext == nil {
-		cfg.UploadContext = contextWithAssetUploadTimeout
-	}
-
-	requestCtx, reqCancel := cfg.RequestContext(ctx)
-	var (
-		set asc.Resource[asc.AppScreenshotSetAttributes]
-		err error
-	)
-	if cfg.DryRun {
-		set, err = findScreenshotSetWithAccess(requestCtx, cfg.Client, cfg.LocalizationID, cfg.DisplayType, cfg.Access)
-	} else {
-		set, err = ensureScreenshotSetWithAccess(requestCtx, cfg.Client, cfg.LocalizationID, cfg.DisplayType, cfg.Access)
-	}
-	reqCancel()
-	if err != nil {
-		return zero, err
-	}
-
-	existingScreenshots := make([]asc.Resource[asc.AppScreenshotAttributes], 0)
-	if (cfg.SkipExisting || cfg.Replace || (!cfg.Replace && len(cfg.Files) > 0)) && set.ID != "" {
-		fetchCtx, fetchCancel := cfg.RequestContext(ctx)
-		existingResp, err := cfg.Client.GetAppScreenshots(fetchCtx, set.ID)
-		fetchCancel()
-		if err != nil {
-			return zero, err
-		}
-		existingScreenshots = existingResp.Data
-	}
-
-	skippedResults := make([]asc.AssetUploadResultItem, 0)
-	files := cfg.Files
-	if cfg.SkipExisting {
-		var filterErr error
-		files, skippedResults, filterErr = filterExistingScreenshotFiles(cfg.Files, existingScreenshots)
-		if filterErr != nil {
-			return zero, filterErr
-		}
-	}
-	files, err = limitScreenshotUploadFilesForExistingSet(files, cfg.MaxScreenshots, existingScreenshots, cfg.Replace, set.ID)
-	if err != nil {
-		return zero, err
-	}
-
-	if cfg.DryRun {
-		results := make([]asc.AssetUploadResultItem, 0, len(skippedResults)+len(files)+len(existingScreenshots))
-		if cfg.Replace {
-			for _, screenshot := range existingScreenshots {
-				results = append(results, asc.AssetUploadResultItem{
-					FileName: screenshot.Attributes.FileName,
-					AssetID:  screenshot.ID,
-					State:    "would-delete",
-				})
-			}
-		}
-		for _, filePath := range files {
-			results = append(results, asc.AssetUploadResultItem{
-				FileName: filepath.Base(filePath),
-				FilePath: filePath,
-				State:    "would-upload",
-			})
-		}
-		results = append(results, skippedResults...)
-		return cfg.BuildResult(cfg.LocalizationID, set, true, results), nil
-	}
-
-	uploadCtx, cancel := cfg.UploadContext(ctx)
-	defer cancel()
-
-	if cfg.Replace {
-		if err := deleteExistingScreenshots(uploadCtx, cfg.Client, existingScreenshots); err != nil {
-			return zero, err
-		}
-	}
-
-	results := make([]asc.AssetUploadResultItem, 0, len(skippedResults)+len(files))
-	if len(files) > 0 {
-		uploadedResults, err := UploadScreenshotsToSet(uploadCtx, cfg.Client, set.ID, files, !cfg.Replace)
-		if err != nil {
-			return zero, err
-		}
-		results = append(results, uploadedResults...)
-	}
-	if cfg.SkipExisting && len(skippedResults) > 0 {
-		if _, err := syncSkippedScreenshotOrder(uploadCtx, cfg.Client, set.ID, cfg.Files, skippedResults, results); err != nil {
-			results = append(skippedResults, results...)
-			return cfg.BuildResult(cfg.LocalizationID, set, false, results), err
-		}
-	}
-	results = append(skippedResults, results...)
-
-	return cfg.BuildResult(cfg.LocalizationID, set, false, results), nil
-}
-
-func deleteExistingScreenshots(ctx context.Context, client *asc.Client, screenshots []asc.Resource[asc.AppScreenshotAttributes]) error {
-	for _, screenshot := range screenshots {
-		if err := client.DeleteAppScreenshot(ctx, screenshot.ID); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func filterExistingScreenshotFiles(files []string, screenshots []asc.Resource[asc.AppScreenshotAttributes]) ([]string, []asc.AssetUploadResultItem, error) {
-	existingByChecksum := make(map[string]asc.Resource[asc.AppScreenshotAttributes], len(screenshots))
-	for _, screenshot := range screenshots {
-		checksum := strings.TrimSpace(screenshot.Attributes.SourceFileChecksum)
-		if checksum == "" {
-			continue
-		}
-		if _, exists := existingByChecksum[checksum]; !exists {
-			existingByChecksum[checksum] = screenshot
-		}
-	}
-
-	filtered := make([]string, 0, len(files))
-	skipped := make([]asc.AssetUploadResultItem, 0)
-	for _, filePath := range files {
-		checksum, err := screenshotFileChecksumFunc(filePath)
-		if err != nil {
-			return nil, nil, err
-		}
-		if existing, exists := existingByChecksum[checksum]; exists {
-			skipped = append(skipped, asc.AssetUploadResultItem{
-				FileName: filepath.Base(filePath),
-				FilePath: filePath,
-				AssetID:  existing.ID,
-				State:    "skipped",
-				Skipped:  true,
-			})
-			continue
-		}
-		filtered = append(filtered, filePath)
-	}
-
-	return filtered, skipped, nil
-}
-
-func syncSkippedScreenshotOrder(ctx context.Context, client *asc.Client, setID string, files []string, skippedResults, uploadedResults []asc.AssetUploadResultItem) ([]string, error) {
-	currentOrder, err := GetOrderedAppScreenshotIDs(ctx, client, setID)
-	if err != nil {
-		return nil, err
-	}
-
-	orderedIDs := orderScreenshotIDsForLocalFiles(currentOrder, files, skippedResults, uploadedResults)
-	if sameScreenshotIDOrder(currentOrder, orderedIDs) {
-		return orderedIDs, nil
-	}
-	return orderedIDs, SetOrderedAppScreenshots(ctx, client, setID, orderedIDs)
-}
-
-func orderScreenshotIDsForLocalFiles(currentOrder []string, files []string, skippedResults, uploadedResults []asc.AssetUploadResultItem) []string {
-	skippedByPath := make(map[string]string, len(skippedResults))
-	for _, item := range skippedResults {
-		if strings.TrimSpace(item.AssetID) == "" {
-			continue
-		}
-		skippedByPath[item.FilePath] = item.AssetID
-	}
-	uploadedByPath := make(map[string]string, len(uploadedResults))
-	for _, item := range uploadedResults {
-		if strings.TrimSpace(item.AssetID) == "" {
-			continue
-		}
-		uploadedByPath[item.FilePath] = item.AssetID
-	}
-
-	orderedIDs := make([]string, 0, len(currentOrder)+len(uploadedResults))
-	seen := make(map[string]struct{}, len(currentOrder)+len(uploadedResults))
-	for _, filePath := range files {
-		id := skippedByPath[filePath]
-		if id == "" {
-			id = uploadedByPath[filePath]
-		}
-		id = strings.TrimSpace(id)
-		if id == "" {
-			continue
-		}
-		if _, exists := seen[id]; exists {
-			continue
-		}
-		seen[id] = struct{}{}
-		orderedIDs = append(orderedIDs, id)
-	}
-	for _, id := range currentOrder {
-		id = strings.TrimSpace(id)
-		if id == "" {
-			continue
-		}
-		if _, exists := seen[id]; exists {
-			continue
-		}
-		seen[id] = struct{}{}
-		orderedIDs = append(orderedIDs, id)
-	}
-
-	return orderedIDs
-}
-
-func sameScreenshotIDOrder(a, b []string) bool {
-	a = normalizeScreenshotIDs(a)
-	b = normalizeScreenshotIDs(b)
-	if len(a) != len(b) {
-		return false
-	}
-	for i := range a {
-		if a[i] != b[i] {
-			return false
-		}
-	}
-	return true
-}
-
-func computeFileChecksum(filePath string) (string, error) {
-	file, err := shared.OpenExistingNoFollow(filePath)
-	if err != nil {
-		return "", err
-	}
-	defer file.Close()
-
-	checksum, err := asc.ComputeChecksumFromReader(file, asc.ChecksumAlgorithmMD5)
-	if err != nil {
-		return "", err
-	}
-	return checksum.Hash, nil
-}
-
-func uploadScreenshotAsset(ctx context.Context, client *asc.Client, setID, filePath string) (asc.AssetUploadResultItem, error) {
-	if err := asc.ValidateImageFile(filePath); err != nil {
-		return asc.AssetUploadResultItem{}, err
-	}
-
-	file, err := shared.OpenExistingNoFollow(filePath)
-	if err != nil {
-		return asc.AssetUploadResultItem{}, err
-	}
-	defer file.Close()
-
-	info, err := file.Stat()
-	if err != nil {
-		return asc.AssetUploadResultItem{}, err
-	}
-
-	checksum, err := asc.ComputeChecksumFromReader(file, asc.ChecksumAlgorithmMD5)
-	if err != nil {
-		return asc.AssetUploadResultItem{}, err
-	}
-
-	created, err := client.CreateAppScreenshot(ctx, setID, info.Name(), info.Size())
-	if err != nil {
-		return asc.AssetUploadResultItem{}, err
-	}
-	if len(created.Data.Attributes.UploadOperations) == 0 {
-		return asc.AssetUploadResultItem{}, fmt.Errorf("no upload operations returned for %q", info.Name())
-	}
-
-	if err := asc.UploadAssetFromFile(ctx, file, info.Size(), created.Data.Attributes.UploadOperations); err != nil {
-		return asc.AssetUploadResultItem{}, err
-	}
-
-	if _, err := client.UpdateAppScreenshot(ctx, created.Data.ID, true, checksum.Hash); err != nil {
-		return asc.AssetUploadResultItem{}, err
-	}
-
-	state, err := waitForScreenshotDelivery(ctx, client, created.Data.ID)
-	if err != nil {
-		return asc.AssetUploadResultItem{}, err
-	}
-
-	return asc.AssetUploadResultItem{
-		FileName: info.Name(),
-		FilePath: filePath,
-		AssetID:  created.Data.ID,
-		State:    state,
-	}, nil
-}
-
-// UploadScreenshotAsset uploads a screenshot file to a set.
-func UploadScreenshotAsset(ctx context.Context, client *asc.Client, setID, filePath string) (asc.AssetUploadResultItem, error) {
-	return uploadScreenshotAsset(ctx, client, setID, filePath)
-}
-
-func waitForScreenshotDelivery(ctx context.Context, client *asc.Client, screenshotID string) (string, error) {
-	return waitForAssetDeliveryState(ctx, screenshotID, func(ctx context.Context) (*asc.AssetDeliveryState, error) {
-		resp, err := client.GetAppScreenshot(ctx, screenshotID)
-		if err != nil {
-			return nil, err
-		}
-		return resp.Data.Attributes.AssetDeliveryState, nil
-	})
 }
